scripts/requirements.txt
python-dotenv>=0.19.0
requests>=2.25.0
opencv-python>=4.5.0
pytesseract>=0.3.8
yt-dlp>=2022.1.21
